SaleController에서 ajax로 통신하는 부분 SaleRestController로 분리, ExceptionHandler 적용, JS코드 리팩토링하려고 함. 가격제시/나눔신청 시 radio 버튼 취소 버튼 만들었음. 판매카테고리 함수 2번씩 호출하는거 수정
