+++ src/main/java/team/cheese/service/sale/SaleService.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import team.cheese.dao.*;
 import team.cheese.dao.MyPage.UserInfoDao;
 import team.cheese.domain.MyPage.SearchCondition;
+import team.cheese.domain.SaleCategoryDto;
 import team.cheese.domain.SaleDto;
 import team.cheese.domain.SaleTagDto;
 import team.cheese.domain.TagDto;
@@ -316,6 +317,16 @@ public class SaleService {
     @Transactional(propagation = Propagation.REQUIRED)
     public void increamentCompleteCnt(String ur_id) throws Exception {
         userInfoDao.incrementCompleteCnt(ur_id);
+    }
+
+    public List<SaleCategoryDto> selectCategory2(String category) throws Exception{
+        List<SaleCategoryDto> saleCategory = saleCategoryDao.selectCategory2(category);
+        return saleCategory;
+    }
+
+    public List<SaleCategoryDto> selectCategory3(String category) throws Exception{
+        List<SaleCategoryDto> saleCategory = saleCategoryDao.selectCategory3(category);
+        return saleCategory;
     }
 }
